:sparkles: Move github api code to local lib
