+++ lib/github.js
@@ -57,3 +57,55 @@ export const fetchRelease = async (githubToken, version = LATEST_RELEASE) => {
   }
 }
 
+async function fetchContributorsOfRepo(orgname, reponame, requestOptions) {
+  return fetch(
+    `https://api.github.com/repos/${orgname}/${reponame}/contributors`,
+    requestOptions
+  );
+}
+
+export async function getContributorsOfAllLinkedRepos(githubToken) {
+  const requestOptions = githubToken
+    ? { headers: { Authorization: `Bearer ${githubToken}` } }
+    : {};
+  const responses = [
+    fetchContributorsOfRepo("lostdesign", "linked", requestOptions),
+    fetchContributorsOfRepo("uselinked", "website", requestOptions),
+    fetchContributorsOfRepo("uselinked", "discord", requestOptions),
+  ];
+  const awaitedResponses = await Promise.all(responses);
+  const isRateLimited = awaitedResponses.reduceRight(
+    (alreadyRatelimited, response) =>
+      alreadyRatelimited || checkRateLimit(response),
+    false
+  );
+  if (isRateLimited) {
+    return { isRateLimited, contributors: [] };
+  }
+  const contributorJsons = await Promise.all(
+    awaitedResponses.map((response) => response.json())
+  );
+  const contributorMap = new Map();
+  contributorJsons.forEach((arrayOfContributors) => {
+    arrayOfContributors.forEach((contributor) => {
+      if (contributor.type !== "User") {
+        // don't count bots and potentially other non-user accounts
+        return;
+      }
+      const existingContributor = contributorMap.get(contributor.login);
+      if (existingContributor) {
+        contributorMap.set(contributor.login, {
+          ...contributor,
+          contributions:
+            existingContributor.contributions + contributor.contributions,
+        });
+      } else {
+        contributorMap.set(contributor.login, contributor);
+      }
+    });
+  });
+  return {
+    isRateLimited,
+    contributors: Array.from(contributorMap.values()),
+  };
+}
