+++ pages/community.vue
@@ -6,60 +6,9 @@
 </template>
 
 <script>
-import { checkRateLimit } from "@/lib/github";
+import { getContributorsOfAllLinkedRepos } from "@/lib/github";
 
-async function fetchContributorsOfRepo(orgname, reponame, requestOptions) {
-  return fetch(
-    `https://api.github.com/repos/${orgname}/${reponame}/contributors`,
-    requestOptions
-  );
-}
 
-async function getContributorsOfAllLinkedRepos(githubToken) {
-  const requestOptions = githubToken
-    ? { headers: { Authorization: `Bearer ${githubToken}` } }
-    : {};
-  const responses = [
-    fetchContributorsOfRepo("lostdesign", "linked", requestOptions),
-    fetchContributorsOfRepo("uselinked", "website", requestOptions),
-    fetchContributorsOfRepo("uselinked", "discord", requestOptions),
-  ];
-  const awaitedResponses = await Promise.all(responses);
-  const isRateLimited = awaitedResponses.reduceRight(
-    (alreadyRatelimited, response) =>
-      alreadyRatelimited || checkRateLimit(response),
-    false
-  );
-  if (isRateLimited) {
-    return { isRateLimited, contributors: [] };
-  }
-  const contributorJsons = await Promise.all(
-    awaitedResponses.map((response) => response.json())
-  );
-  const contributorMap = new Map();
-  contributorJsons.forEach((arrayOfContributors) => {
-    arrayOfContributors.forEach((contributor) => {
-      if (contributor.type !== "User") {
-        // don't count bots and potentially other non-user accounts
-        return;
-      }
-      const existingContributor = contributorMap.get(contributor.login);
-      if (existingContributor) {
-        contributorMap.set(contributor.login, {
-          ...contributor,
-          contributions:
-            existingContributor.contributions + contributor.contributions,
-        });
-      } else {
-        contributorMap.set(contributor.login, contributor);
-      }
-    });
-  });
-  return {
-    isRateLimited,
-    contributors: Array.from(contributorMap.values()),
-  };
-}
 
 export default {
   async asyncData({ $config: { githubToken } }) {
